hostlayout clean up
