+++ frontend/src/components/HostLayout.jsx
@@ -1,58 +1,12 @@
-import React, { useEffect, useRef, useState } from 'react';
+import React from 'react';
 import { Route, Routes } from 'react-router-dom';
-import { SockJsClientDefaults } from '../services/SockJsClientDefaults';
-import FinalResults from '../views/HostViews/FinalResults';
-import MainQuiz from '../views/HostViews/MainQuiz';
-import QuizRoom from '../views/HostViews/QuizRoom';
-import TestView from '../views/TestView';
 import NavigationBar from './NavigationBar/NavigationBar';
+import HostLayoutConnected from './HostLayoutConnected';
+import HostLayoutNotConnected from './HostLayoutNotConnected';
 
-function HostLayout() {
-  const [topics, setTopics] = useState([]);
-  const [connected, setConnected] = useState(false);
-
-  const [handleMessage, setHandleMessage] = useState({
-    fn: (msg) => console.log(msg),
-  });
-  const [handleConnect, setHandleConnect] = useState({
-    fn: () => {
-      console.log('connected');
-      setConnected(true);
-    },
-  });
-
-  const [handleDisconnect, setHandleDisconnect] = useState({
-    fn: () => console.log('disconnected'),
-  });
-
-  const client = useRef(null);
-
-  useEffect(() => {
-    setConnected(client.current?.state.connected);
-  }, [client.current?.state]);
-
-  const props = {
-    client,
-    topics,
-    setTopics,
-    handleMessage,
-    setHandleMessage,
-    handleConnect,
-    setHandleConnect,
-    handleDisconnect,
-    setHandleDisconnect,
-    connected,
-  };
-
+function HostLayout(props) {
   return (
     <>
-      <SockJsClientDefaults
-        topics={topics}
-        onConnect={handleConnect.fn}
-        onDisconnect={handleDisconnect.fn}
-        onMessage={handleMessage.fn}
-        ref={client}
-      />
       <NavigationBar
         buttons={[
           { to: '/', text: 'Home' },
@@ -62,10 +16,8 @@ function HostLayout() {
         loggedIn={true}
       />
       <Routes>
-        <Route path=":id" element={<QuizRoom {...props} />} />
-        <Route path=":id/quiz-room/:pin" element={<MainQuiz {...props} />} />
-        <Route path=":id/finalresults/:pin" element={<FinalResults {...props} />} />
-        <Route path="test" element={<TestView {...props} />} />
+        <Route path="/" element={<HostLayoutNotConnected {...props} />} />
+        <Route path="/*" element={<HostLayoutConnected {...props} />} />
       </Routes>
     </>
   );
